+++ src/main/java/edu/ie3/util/copy/FastByteArrayInputStream.java
@@ -11,7 +11,7 @@ import java.io.InputStream;
  * ByteArrayInputStream implementation that does not synchronize methods.
  *
  * @version 0.1
- * @author: hiry
+ * @author hiry
  * @since 13.06.2018
  */
 public class FastByteArrayInputStream extends InputStream {
@@ -29,14 +29,17 @@ public class FastByteArrayInputStream extends InputStream {
     this.count = count;
   }
 
+  @Override
   public final int available() {
     return count - pos;
   }
 
+  @Override
   public final int read() {
     return (pos < count) ? (buf[pos++] & 0xff) : -1;
   }
 
+  @Override
   public final int read(byte[] b, int off, int len) {
     if (pos >= count) return -1;
 
@@ -47,8 +50,9 @@ public class FastByteArrayInputStream extends InputStream {
     return len;
   }
 
+  @Override
   public final long skip(long n) {
-    if ((pos + n) > count) n = count - pos;
+    if ((pos + n) > count) n = (long) count - pos;
     if (n < 0) return 0;
     pos += n;
     return n;